requirements.txt
fastapi==0.99.1
uvicorn==0.30.3
opencv-python==4.9.0.80
numpy==1.26.2
Pillow==10.2.0
torch==2.2.2
facenet-pytorch==2.6.0
deepface==0.0.92
torchvision==0.17.2
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
fastapi-jwt-auth==0.5.0
pydantic==1.10.17